# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ShopApp E-Commerce API",
        "description": "REST API for the ShopApp e-commerce platform. Provides endpoints for authentication, product browsing, cart management, checkout, and order history.",
        "contact": {
            "name": "API Support",
            "email": "support@shopapp.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/checkout": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Place an order from the active cart",
                "operationId": "d480c2234adc7930df2e78be75f98b93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shipping_address",
                                    "payment_method"
                                ],
                                "properties": {
                                    "shipping_address": {
                                        "type": "string",
                                        "example": "123 Main Street, City, State 12345"
                                    },
                                    "billing_address": {
                                        "type": "string",
                                        "example": "456 Billing Ave, City, State 67890",
                                        "nullable": true
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "credit_card",
                                        "enum": [
                                            "credit_card",
                                            "debit_card",
                                            "paypal"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order placed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order placed successfully"
                                        },
                                        "order": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cart is empty"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/checkout/pay/{orderId}": {
            "get": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Process payment for an order (simulated)",
                "description": "Simulates a payment gateway. Has ~80% success rate. Updates order status to 'paid' on success.",
                "operationId": "cbafb3563a6b9784c775bd286193e94f",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "Order ID to process payment for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment successful"
                                        },
                                        "order": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List all active products",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by product name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "headphones"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 22
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a single product by ID",
                "operationId": "b5099368868dbe7900acd129345eb148",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Wireless Bluetooth Headphones"
                    },
                    "slug": {
                        "type": "string",
                        "example": "wireless-bluetooth-headphones"
                    },
                    "description": {
                        "type": "string",
                        "example": "High quality headphones"
                    },
                    "price": {
                        "type": "string",
                        "example": "79.99"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 50
                    },
                    "image_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "category": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "ProductDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Wireless Bluetooth Headphones"
                    },
                    "slug": {
                        "type": "string",
                        "example": "wireless-bluetooth-headphones"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string",
                        "example": "79.99"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 50
                    },
                    "image_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "string",
                        "example": "159.98"
                    },
                    "status": {
                        "type": "string",
                        "example": "paid",
                        "enum": [
                            "pending",
                            "paid",
                            "failed",
                            "refunded"
                        ]
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "credit_card"
                    },
                    "shipping_address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "billing_address": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Wireless Bluetooth Headphones"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "type": "string",
                        "example": "79.99"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token obtained from /login or /register",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}